Add filter by time period to comment list
A new button is now displayed in comment list view for creators and participants. A click on the button will open
a menu with which the user can select a time period for the displayed comments. Sorting and filtering actions will
applied to the selected period. Displaying all existing comments is the default setting.