Add new time period filter 'from now'
Users are now able to filter the comments from now. If an user has selected this option all comments disappear and only comments created from now will be displayed.
If the user selects another time period or the 'from now' option again, the period will reset so that the comments will disappear again.