Improve the performance of theme cloud administration
The keyword update function has a runtime that is quadratic
to the number of keywords, since the function examines all
keywords and has to search all previous keywords each time.
This commit ensures that this function is called only once
in ngOnInit instead of three times previously, as the
additional calls are redundant, reducing the administration
call time to one-third.

Closes #119